# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/plans": {
      "post": {
        "description": "Cria um novo plano no sistema. Requer permissões de administrador.",
        "operationId": "PlansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plano criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "409": {
            "description": "Plano com este slug já existe"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar novo plano",
        "tags": [
          "📋 Planos"
        ]
      },
      "get": {
        "description": "Retorna lista de planos disponíveis. Por padrão, retorna apenas planos ativos.",
        "operationId": "PlansController_findAll",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Incluir planos inativos na listagem",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de planos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Listar todos os planos",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/{id}": {
      "get": {
        "description": "Retorna detalhes de um plano específico pelo ID.",
        "operationId": "PlansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plano encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "summary": "Buscar plano por ID",
        "tags": [
          "📋 Planos"
        ]
      },
      "patch": {
        "description": "Atualiza dados de um plano existente. Requer permissões de administrador.",
        "operationId": "PlansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plano atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar plano",
        "tags": [
          "📋 Planos"
        ]
      },
      "delete": {
        "description": "Remove um plano do sistema (soft delete). Requer permissões de administrador.",
        "operationId": "PlansController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Plano removido com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover plano",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/slug/{slug}": {
      "get": {
        "description": "Retorna detalhes de um plano específico pelo slug.",
        "operationId": "PlansController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Slug único do plano",
            "schema": {
              "example": "premium-plan",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plano encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "summary": "Buscar plano por slug",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/type/{type}": {
      "get": {
        "description": "Retorna lista de planos filtrados por tipo.",
        "operationId": "PlansController_findByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Tipo do plano",
            "schema": {
              "enum": [
                "free",
                "basic",
                "premium",
                "enterprise"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Planos encontrados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar planos por tipo",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/admin/stats": {
      "get": {
        "description": "Retorna estatísticas gerais dos planos cadastrados. Requer permissões de administrador.",
        "operationId": "PlansController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estatísticas retornadas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "number",
                      "example": 10
                    },
                    "active": {
                      "type": "number",
                      "example": 8
                    },
                    "inactive": {
                      "type": "number",
                      "example": 2
                    },
                    "featured": {
                      "type": "number",
                      "example": 3
                    },
                    "byType": {
                      "type": "object",
                      "example": {
                        "free": 2,
                        "premium": 5,
                        "enterprise": 3
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter estatísticas dos planos",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/{id}/toggle-active": {
      "patch": {
        "description": "Alterna o status ativo de um plano. Requer permissões de administrador.",
        "operationId": "PlansController_toggleActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status do plano alterado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ativar/Desativar plano",
        "tags": [
          "📋 Planos"
        ]
      }
    },
    "/v1/plans/{id}/featured": {
      "patch": {
        "description": "Define se um plano é destacado na listagem. Requer permissões de administrador.",
        "operationId": "PlansController_setFeatured",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "featured",
            "required": true,
            "in": "query",
            "description": "Se o plano deve ser destacado",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status destacado alterado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Definir plano como destacado",
        "tags": [
          "📋 Planos"
        ]
      }
    },
...
    "/v1/plans/{planId}/prices/{priceId}": {
      "patch": {
        "description": "Atualiza um preço específico de um plano. Requer permissões de administrador.",
        "operationId": "PlansController_updatePlanPrice",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "priceId",
            "required": true,
            "in": "path",
            "description": "ID único do preço",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanPriceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preço atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "planId": {
                      "type": "string",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "period": {
                      "type": "string",
                      "example": "monthly"
                    },
                    "price": {
                      "type": "number",
                      "example": 29.99
                    },
                    "currency": {
                      "type": "string",
                      "example": "BRL"
                    },
                    "isActive": {
                      "type": "boolean",
                      "example": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano ou preço não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar preço de um plano",
        "tags": [
          "📋 Planos"
        ]
      },
      "delete": {
        "description": "Remove um preço específico de um plano. Requer permissões de administrador.",
        "operationId": "PlansController_deletePlanPrice",
        "parameters": [
          {
            "name": "planId",
            "required": true,
            "in": "path",
            "description": "ID único do plano",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "name": "priceId",
            "required": true,
            "in": "path",
            "description": "ID único do preço",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Preço removido com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "403": {
            "description": "Acesso negado - permissões insuficientes"
          },
          "404": {
            "description": "Plano ou preço não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remover preço de um plano",
        "tags": [
          "📋 Planos"
        ]
      }
    },
...
  },
  "info": {
    "title": "🧠 NeuralContent API",
    "description": "\n## API completa para plataforma de geração de conteúdo com IA\n\n### 🔐 Autenticação\nEsta API utiliza **JWT Bearer tokens**. Para autenticar suas requisições:\n```\nAuthorization: Bearer {seu_jwt_token}\n```\n\n### 📊 Rate Limiting\n- **Usuários Free:** 100 req/min\n- **Usuários Pro:** 500 req/min  \n- **Usuários Enterprise:** 1000 req/min\n- **Administradores:** Sem limite\n\n### 🔄 Versionamento\n- **Versão Atual:** v1.0.0\n- **URL Base:** `/v1`\n- **Suporte:** Versionamento semântico (SemVer)\n\n### 📈 Funcionalidades Principais\n- **Autenticação e Autorização** (JWT + Roles)\n- **Gestão de Usuários** (CRUD + Perfis)\n- **Sistema de Planos** (Free, Pro, Enterprise)\n- **Gestão de Créditos** (Compra, Consumo, Histórico)\n- **Processamento de Pagamentos** (Stripe Integration)\n- **Monitoramento e Logs** (Health Checks + Analytics)\n\n### 🛡️ Segurança\n- Criptografia de senhas (bcrypt)\n- Validação de entrada (class-validator)\n- Proteção CORS configurável\n- Headers de segurança (Helmet)\n- Rate limiting por usuário\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "NeuralContent Support",
      "url": "https://neuralcontent.com/support",
      "email": "api-support@neuralcontent.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://neuralcontent.com/terms"
  },
...
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT token obtido via /auth/login. Formato: Bearer {token}",
        "in": "header"
      },
      "ApiKey-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para integração de sistemas externos"
      }
    },
...
  }
}